Try building the library so it works with react
CRA does not have support for class properties so direct import will fail. We need to
compile the library to something importable by CRA apps.

(change) Compile the library code before releasing. We’re using experimental
features (class properties) so we need to copile it before sharing with the rest
of the world.