Remove validation and external dependencies
(change) Remove TypedModel.validate(). It was the only place jsonschema was used and
we do not need it as an unnecessary dependency.

(change) Reimplement date formatter to not use date-fns. Same as the above, this is a
very simple case and no need to pull an external library just for it.

The main reason for avoiding external deps is the fact that since we’re using
experimental JS features (class properties) we need to compile the library before
it can be used inside create-react-app project. We don’t want to bundle anything to
not increase the size unnecesarily.